♻️  convert adjustFontSize into an util function
